splitting things up
