+++ haskelloids.cabal
@@ -53,7 +53,9 @@ executable haskelloids
   ghc-options:         -Wall
   
   -- Modules included in this executable, other than Main.
-  -- other-modules:       
+  other-modules:      InGame
+                    , Types 
+                    , Commons
   
   -- LANGUAGE extensions used by modules in this package.
   -- other-extensions:
